% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19_bins}
\alias{hg19_bins}
\title{A dataframe containing location information for 1MB-size bins within the hg19 genome.
This dataset is generated based on the accessed cytoband data and used for
CNV frequency calculation from local pgxseg files.}
\format{
An object of class \code{data.frame} with 3114 rows and 6 columns.
}
\usage{
hg19_bins
}
\value{
genomic bin locations of hg19 genome
}
\description{
A dataframe containing location information for 1MB-size bins within the hg19 genome.
This dataset is generated based on the accessed cytoband data and used for
CNV frequency calculation from local pgxseg files.
}
\keyword{datasets}
