% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38}
\alias{hg38}
\title{A dataframe containing cytoband annotation details extracted from the hg38 gennome.}
\format{
An object of class \code{data.frame} with 862 rows and 5 columns.
}
\source{
\url{http://hgdownload.cse.ucsc.edu/goldenpath/hg38/database/cytoBand.txt.gz}
}
\usage{
hg38
}
\description{
A dataframe containing cytoband annotation details extracted from the hg38 gennome.
}
\keyword{datasets}
