% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_bins}
\alias{hg38_bins}
\title{A dataframe containing location information for 1MB-size bins within the hg38 genome.
This dataset is generated based on the accessed cytoband data and used for
CNV frequency calculation from local pgxseg files.}
\format{
An object of class \code{data.frame} with 3106 rows and 6 columns.
}
\usage{
hg38_bins
}
\description{
A dataframe containing location information for 1MB-size bins within the hg38 genome.
This dataset is generated based on the accessed cytoband data and used for
CNV frequency calculation from local pgxseg files.
}
\keyword{datasets}
