% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxFilter.R
\name{pgxFilter}
\alias{pgxFilter}
\title{Query available filters}
\usage{
pgxFilter(
  prefix = NULL,
  return_all_prefix = FALSE,
  domain = "http://progenetix.org",
  entry_point = "beacon"
)
}
\arguments{
\item{prefix}{A string specifying the prefix of filters, such as 'NCIT'. Default is NULL, which
means that all available filters will be returned. When specified, it returns all filters with the specified prefix.}

\item{return_all_prefix}{A logical value determining whether to return all valid prefixes of filters used in database.
If TRUE, the \code{prefix} parameter will be ignored. Default is FALSE.}

\item{domain}{A string specifying the domain of the query data resource. Default is "http://progenetix.org".}

\item{entry_point}{A string specifying the entry point of the Beacon v2 API. Default is "beacon", resulting in the endpoint being "http://progenetix.org/beacon".}
}
\value{
Filter terms used in the data resource that you query.
}
\description{
This function retrieves available filters in the Progenetix database via the Beacon v2 API.
}
\examples{
pgxFilter(prefix = "NCIT")
}
