% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxFreqselect.R
\name{pgxFreqselect}
\alias{pgxFreqselect}
\title{Return object with selected ids}
\usage{
pgxFreqselect(data, id)
}
\arguments{
\item{data}{CNV frequency data returned by \code{pgxLoader} function}

\item{id}{A single or a comma-concatenated list of identifiers such as c("NCIT:C7376","icdom-98353")}
}
\value{
A list object with the same format of input but only containing selected id
}
\description{
This function helps select ids of interests from input object.
}
