% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxSegprocess.R
\name{pgxSegprocess}
\alias{pgxSegprocess}
\title{Extract, analyse and visualize "pgxseg" files}
\usage{
pgxSegprocess(
  file,
  group_id = "group_id",
  show_KM_plot = FALSE,
  return_metadata = FALSE,
  return_seg = FALSE,
  return_frequency = FALSE,
  assembly = "hg38",
  ...
)
}
\arguments{
\item{file}{A string specifying the path and name of the "pgxseg" file where the data is to be read.}

\item{group_id}{A string specifying which id is used for grouping in KM plot or CNV frequency calculation. Default is "group_id".}

\item{show_KM_plot}{A logical value determining whether to return the Kaplan-Meier plot based on metadata. Default is FALSE.}

\item{return_metadata}{A logical value determining whether to return metadata. Default is FALSE.}

\item{return_seg}{A logical value determining whether to return segment data. Default is FALSE.}

\item{return_frequency}{A logical value determining whether to return CNV frequency data. The frequency calculation is based on segments in segment data and specified group id in metadata. Default is FALSE.}

\item{assembly}{A string specifying which genome assembly version should be applied to CNV frequency calculation and plotting. Allowed options are "hg19", "hg38". Default is "hg38".}

\item{...}{Other parameters relevant to KM plot. These include \code{pval}, \code{pval.coord}, \code{pval.method}, \code{conf.int}, \code{linetype}, and \code{palette} (see ggsurvplot from survminer)}
}
\value{
Segments data, CNV frequency object, meta data or KM plots from local "pgxseg" files
}
\description{
This function extracts segments, CNV frequency, and metadata from local "pgxseg" files and
supports survival data visualization
}
\examples{
info <- pgxSegprocess(file=file_name,show_KM_plot = TRUE, return_seg = TRUE, return_metadata = TRUE, return_frequency = TRUE)
}
