% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19}
\alias{hg19}
\title{A dataframe containing cytoband annotation details extracted from the hg19 gennome.
It is used for CNV frequency visualization.}
\format{
An object of class \code{data.frame} with 862 rows and 5 columns.
}
\source{
\url{http://hgdownload.cse.ucsc.edu/goldenpath/hg19/database/cytoBand.txt.gz}
}
\usage{
hg19
}
\value{
cytoband of hg19 genome
}
\description{
A dataframe containing cytoband annotation details extracted from the hg19 gennome.
It is used for CNV frequency visualization.
}
\keyword{datasets}
