% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxCount.R
\name{pgxCount}
\alias{pgxCount}
\title{Count samples in one collation of a given id}
\usage{
pgxCount(group_id = NULL)
}
\arguments{
\item{group_id}{A single or a comma-concatenated list of identifiers such as c("NCIT:C7376","icdom-98353")}
}
\value{
Count of samples in the given id
}
\description{
This function returns the number of samples for every group id in
Progenetix database.
}
