% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxFilter.R
\name{pgxFilter}
\alias{pgxFilter}
\title{Query available filters}
\usage{
pgxFilter(
  prefix = NULL,
  return_all_prefix = FALSE,
  domain = "http://progenetix.org",
  dataset = "Progenetix"
)
}
\arguments{
\item{prefix}{A string specifying the prefix of filters, such as 'NCIT' and 'PMID'. Default is NULL, which
means that all available filters will be returned. When specified, it returns all filters with the specified prefix.}

\item{return_all_prefix}{A logical value determining whether to return all valid prefixes of filters used in Progenetix.
If TRUE, the \code{prefix} parameter will be ignored. Default is FALSE.}

\item{domain}{A string specifying the domain of the Progenetix database. Default is "http://progenetix.org".}

\item{dataset}{A string specifying the dataset to query. Default is "Progenetix". Other available options are "cancercelllines".}
}
\value{
filter terms used in Progenetix.
}
\description{
This function retrieves available filters in the Progenetix database.
}
\examples{
pgxFilter(prefix = "NCIT")
}
