% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxFreqplot.R
\name{pgxFreqplot}
\alias{pgxFreqplot}
\title{Plot CNV frequency data}
\usage{
pgxFreqplot(
  data,
  chrom = NULL,
  layout = c(1, 1),
  filters = NULL,
  circos = FALSE,
  highlight = NULL,
  assembly = "hg38"
)
}
\arguments{
\item{data}{Frequency object returned by \code{pgxLoader} function or individual frequency matrix with 'pgxfreq' format.
In the returned object by \code{pgxLoader}, the frequency matrices in \code{data} slot must be stored as 'pgxfreq' format,
which can be specified by \code{output} parameter of \code{pgxLoader} function.}

\item{chrom}{A vector with chromosomes to be plotted. If NULL, return the plot
by genome. If specified the frequencies are plotted with one panel for each
chromosome. Default is NULL.}

\item{layout}{Number of columns and rows in plot. Only used in plot by chromosome.
Default is c(1,1).}

\item{filters}{Index or string value to indicate which filter to be plotted, such as 1
(the first filters in \code{data} slot of object ) or 'NCIT:C4038' (specific filter name). The length of filters
is limited to one if the parameter \code{circos} is False. Default is 1.}

\item{circos}{A logical value to indicate if return a circos plot. If TRUE, it
can return a circos plot with multiple filters for display and comparison.
Default is FALSE.}

\item{highlight}{Indices of genomic bins to be highlighted with red color.}

\item{assembly}{A string specifying which genome assembly version should be applied to CNV frequency plotting.
Allowed options are "hg19", "hg38". Default is "hg38" (genome version used in Progenetix).}
}
\value{
The binned CNV frequency plot
}
\description{
Thie function plots the frequency of deletions and amplifications
}
\examples{
## get frequency data
freq <- pgxLoader(type="frequency", output ='pgxfreq', filters="NCIT:C3512")
## visualize
pgxFreqplot(freq)
}
