% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxFreqselect.R
\name{pgxFreqselect}
\alias{pgxFreqselect}
\title{Return object with selected filters}
\usage{
pgxFreqselect(data, filters)
}
\arguments{
\item{data}{CNV frequency data returned by \code{pgxLoader} function}

\item{filters}{A single or a comma-concatenated list of identifiers such as c("NCIT:C7376","icdom-98353")}
}
\value{
A list object with the same format of input but only containing selected filters
}
\description{
This function helps select filters of interests from input object.
}
\examples{
total_freq <- pgxLoader(type="frequency", output ='pgxmatrix', filters=c("NCIT:C3512","NCIT:C3059","NCIT:C3716","NCIT:C4917"))
subset_freq <-  pgxFreqselect(data=total_freq, filters = c("NCIT:C3512","NCIT:C3059"))
}
