% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgxLoader.R
\name{pgxLoader}
\alias{pgxLoader}
\title{Load data from Progenetix database}
\usage{
pgxLoader(
  type = NULL,
  output = NULL,
  group_id = NULL,
  codematches = FALSE,
  biosample_id = NULL,
  save_file = FALSE,
  filename = NULL
)
}
\arguments{
\item{type}{A string specifying output data type. Avaiable options are "biosample",
"variant" or "frequency".}

\item{output}{A string specifying output file format. Avaiable options are NULL,
'pgxseg' or 'seg' when the paramter \code{type} is "variant"; 'pgxseg' or 'pgxmatrix'
when the paramter \code{type} is "frequency".}

\item{group_id}{A single or a comma-concatenated list of identifiers for cancer type,
literature, and cohorts such as c("NCIT:C7376","icdom-98353","PMID:22824167", "pgxcohort-TCGAcancers")}

\item{codematches}{A logical value determining whether to exclude samples
from child terms of specified group id. If FALSE, retrieved samples include child
terms of specified group id. Default is FALSE.}

\item{biosample_id}{A single or a comma-concatenated list of identifiers used
in Progenetix database for identifying biosamples}

\item{save_file}{A logical value determining whether to save the variant data as file
instead of direct return. Only used when the paramter \code{type} is "variant". Default is FALSE.}

\item{filename}{A string specifying the path and name of the file to be saved.
Only used if the parameter \code{save_file} is TRUE. Default is "variants.seg/pgxseg"
in current work directory.}
}
\value{
Data from Progenetix database
}
\description{
This function loads various data from \code{Progenetix} database.
}
